Eliminate __init__.py under any test directory; Relocate utilities.
