+++ nucypher/utilities/simulate.py
@@ -1,0 +1,12 @@
+import random
+
+from constant_sorrow import constants
+from twisted.internet import protocol
+
+
+class UrsulaProcessProtocol(protocol.ProcessProtocol):
+    pass
+
+
+class UrsulaStakingProtocol(UrsulaProcessProtocol):
+    pass
